Feat: refactorizacion de las carpetas de almacenamiento
