+++ exam-craft-extension-frontend/src/screens/storage/GenerationCodeScreen.tsx
