+++ exam-craft-extension-frontend/src/screens/storage/css/FoldersGridScreen.css
@@ -1,0 +1,32 @@
+.empty-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+    text-align: center;
+    color: #7a5c4f;
+    font-size: 16px;
+    font-weight: 500;
+    margin-top: 20px;
+    width: 100%;
+    background-color: #e8d5c4;
+    border-radius: 20px;
+    padding: 40px 60px;
+    box-shadow: inset 0 2px 8px rgba(0,0,0,0.08);
+}
+
+.empty-subtext {
+    font-size: 15px;
+    color: #a07060;
+    margin-top: 4px;
+}
+.empty-subtext {
+    font-size: 15px;
+    margin-top: 8px;
+}
+
+.card-count {
+    font-size: 15px;
+    color: #000000;
+    margin-top: 5px;
+}
